requirements-test.txt
# Dependências para testes
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-xdist==3.3.1

# Selenium para testes end-to-end
selenium==4.15.2
webdriver-manager==4.0.1

# Cobertura de código
coverage==7.3.2

# Factory Boy para criação de dados de teste
factory-boy==3.3.0

# Faker para dados fictícios
Faker==20.1.0

# Django Debug Toolbar para desenvolvimento
django-debug-toolbar==4.2.0

# Testes de performance
locust==2.17.0

# Mock para testes
mock==5.1.0

# FreezeGun para testes de data/hora
freezegun==1.2.2
